% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installShinyParallel.R
\name{installShinyParallel}
\alias{installShinyParallel}
\title{Installs a multi-session Shiny app in a server}
\usage{
installShinyParallel(appDir = getwd(), appName = basename(appDir),
  max.sessions = getOption("shinyParallel.max.sessions", 20L),
  users.per.session = getOption("shinyParallel.users.per.session", Inf),
  shinyServerPath = "/srv/shiny-server/")
}
\arguments{
\item{appDir}{The application to run. Should be one of the following:
\itemize{
\item A directory containing \code{server.R}, plus, either \code{ui.R} or
 a \code{www} directory that contains the file \code{index.html}.
\item A directory containing \code{app.R}.
}}

\item{appName}{Name of the app (path to access it on the server).}

\item{max.sessions}{Number of sessions to use. Defaults to the
\code{shinyParallel.max.sessions} option, is set, or \code{2L} if not.}

\item{users.per.session}{Maximum number of admited users per each session.
Defaults to the
\code{shinyParallel.users.per.session} option, is set, or \code{Inf} if
not.}

\item{shinyServerPath}{Path where shiny-server apps are installed by default.}
}
\description{
Installs a Shiny app in a Shiny server, with the multi-session feature
enabled.
It will run in \code{max.sessions}, each with the Shiny app working.
So, comunication between users is limited, if this needs to be done, then
save and load data on hard disk (or use RStudio server pro).
}
\examples{
\dontrun{
# If we have a Shiny app at '~/myShinyApp', i.e., we can test our app by:
# shinyParallel::runApp('~/myShinyApp');

# then we can install the app by typing
shinyParallel::installShinyParallel('~/myShinyApp');
}
}
